+++ Sources/Sentry/SentryFileIOTrackingIntegration.m
@@ -1,7 +1,6 @@
 #import "SentryFileIOTrackingIntegration.h"
 #import "SentryLog.h"
 #import "SentryNSDataSwizzling.h"
-#import "SentryNSFileManagerSwizzling.h"
 #import "SentryOptions.h"
 
 @implementation SentryFileIOTrackingIntegration
@@ -13,7 +12,6 @@
     }
 
     [SentryNSDataSwizzling.shared startWithOptions:options];
-    [SentryNSFileManagerSwizzling.shared startWithOptions:options];
 
     return YES;
 }
@@ -27,7 +25,6 @@
 - (void)uninstall
 {
     [SentryNSDataSwizzling.shared stop];
-    [SentryNSFileManagerSwizzling.shared stop];
 }
 
 @end