+++ Sources/Sentry/include/SentryNSDataTracker.h
@@ -8,7 +8,7 @@ static NSString *const SENTRY_FILE_READ_OPERATION = @"file.read";
 
 @class SentryThreadInspector, SentryNSProcessInfoWrapper;
 
-@interface SentryFileIOTracker : NSObject
+@interface SentryNSDataTracker : NSObject
 SENTRY_NO_INIT
 
 - (instancetype)initWithThreadInspector:(SentryThreadInspector *)threadInspector
@@ -58,16 +58,6 @@ SENTRY_NO_INIT
                                     error:(NSError **)error
                                    method:(NSData *_Nullable (^)(
                                               NSURL *, NSDataReadingOptions, NSError **))method;
-
-/**
- * Measure NSFileManager 'createFileAtPath:contents:attributes::' method.
- */
-- (BOOL)measureNSFileManagerCreateFileAtPath:(NSString *)path
-                                        data:(NSData *)data
-                                  attributes:(NSDictionary<NSFileAttributeKey, id> *)attributes
-                                      method:(BOOL (^)(NSString *, NSData *,
-                                                 NSDictionary<NSFileAttributeKey, id> *))method;
-
 @end
 
 NS_ASSUME_NONNULL_END